refactor: Update imports in teacher page and evaluations page
