+++ src/modules/teacher/components/index.ts
@@ -1,0 +1,7 @@
+import TeacherTable from './table/TeacherTable';
+export * from './aside/AsideCard';
+export * from './header-search/HeaderSearch';
+export * from './search-input/SearchInput';
+export * from './SelectOptions'
+
+export {TeacherTable}
